# openapi.yaml
openapi: 3.0.1
info:
  title: gpt-db API
  version: 1.0.0
servers:
  - url: https://gpt-db.vercel.app
    description: Production server
paths:
...
  /food/stock:
    get:
      summary: List stock
      operationId: getFoodStock
      tags:
        - food
      parameters:
        - in: query
          name: view
          required: false
          schema:
            type: string
            enum: [aggregate, items]
      responses:
        '200':
          description: Successful response
    post:
      summary: Add stock units
      operationId: addFoodStock
      tags:
        - food
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [upc, quantity]
                    properties:
                      upc:
                        type: string
                        description: Barcode for the product
                      quantity:
                        type: integer
                        minimum: 1
                      name:
                        type: string
                        description: Optional product name to seed/update catalog
                      tags:
                        type: array
                        items:
                          type: string
                        description: Optional tags to seed/update catalog
                      ingredients:
                        type: array
                        items:
                          type: string
                        description: Optional ingredients to seed/update catalog
                      nutrition:
                        type: object
                        description: Optional per-unit nutrition facts to seed/update catalog
      responses:
        '201':
          description: Stock updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  upserted_uuids:
                    type: array
                    items:
                      type: string
                  count:
                    type: integer
  /food/stock/consume:
    post:
      summary: Consume stock and log nutrition
      operationId: consumeFoodStock
      tags:
        - food
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [upc]
              properties:
                upc:
                  type: string
                units:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Stock consumed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  remaining:
                    type: integer
  /food/stock/remove:
    post:
      summary: Remove stock without log
      operationId: removeFoodStock
      tags:
        - food
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [upc, units, reason]
              properties:
                upc:
                  type: string
                units:
                  type: integer
                  minimum: 1
                reason:
                  type: string
      responses:
        '200':
          description: Stock removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  remaining:
                    type: integer
  /food/stock/{stock_uuid}:
    delete:
      summary: Delete specific stock row
      operationId: deleteStockRow
      tags:
        - food
      parameters:
        - in: path
          name: stock_uuid
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stock row deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  deleted:
                    type: boolean
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - ApiKeyAuth: []